投放系统Service

+++ watermelon-ad-service/ad-sponsor/src/main/java/com/watermelon/ad/entity/unit_condition/AdUnitDistrict.java
@@ -20,7 +20,7 @@ public class AdUnitDistrict {
 
     @Basic
     @Column(name = "unit_id", nullable = false)
-    private Integer unitId;
+    private Long unitId;
 
     @Basic
     @Column(name = "province", nullable = false)
@@ -30,7 +30,7 @@ public class AdUnitDistrict {
     @Column(name = "city", nullable = false)
     private String city;
 
-    public AdUnitDistrict(Integer unitId, String province, String city) {
+    public AdUnitDistrict(Long unitId, String province, String city) {
         this.unitId = unitId;
         this.province = province;
         this.city = city;
